update source translate vietnamese

+++ common/lib/xmodule/xmodule/modulestore/mongo/draft.py
@@ -8,12 +8,11 @@ and otherwise returns i4x://org/course/cat/name).
 
 from datetime import datetime
 
-from xmodule.exceptions import InvalidVersionError
 from xmodule.modulestore import Location, namedtuple_to_son
-from xmodule.modulestore.exceptions import ItemNotFoundError, DuplicateItemError
+from xmodule.modulestore.exceptions import ItemNotFoundError
 from xmodule.modulestore.inheritance import own_metadata
-from xmodule.modulestore.mongo.base import location_to_query, get_course_id_no_run, MongoModuleStore
-import pymongo
+from xmodule.exceptions import InvalidVersionError
+from xmodule.modulestore.mongo.base import MongoModuleStore
 from pytz import UTC
 
 DRAFT = 'draft'
@@ -93,21 +92,6 @@ class DraftModuleStore(MongoModuleStore):
         except ItemNotFoundError:
             return wrap_draft(super(DraftModuleStore, self).get_instance(course_id, location, depth=depth))
 
-    def create_xmodule(self, location, definition_data=None, metadata=None, system=None):
-        """
-        Create the new xmodule but don't save it. Returns the new module with a draft locator
-
-        :param location: a Location--must have a category
-        :param definition_data: can be empty. The initial definition_data for the kvs
-        :param metadata: can be empty, the initial metadata for the kvs
-        :param system: if you already have an xmodule from the course, the xmodule.system value
-        """
-        draft_loc = as_draft(location)
-        if draft_loc.category in DIRECT_ONLY_CATEGORIES:
-            raise InvalidVersionError(location)
-        return super(DraftModuleStore, self).create_xmodule(draft_loc, definition_data, metadata, system)
-
-
     def get_items(self, location, course_id=None, depth=0):
         """
         Returns a list of XModuleDescriptor instances for the items
@@ -135,26 +119,14 @@ class DraftModuleStore(MongoModuleStore):
         ]
         return [wrap_draft(item) for item in draft_items + non_draft_items]
 
-    def convert_to_draft(self, source_location):
+    def clone_item(self, source, location):
         """
-        Create a copy of the source and mark its revision as draft.
-
-        :param source: the location of the source (its revision must be None)
+        Clone a new item that is a copy of the item at the location `source`
+        and writes it to `location`
         """
-        original = self.collection.find_one(location_to_query(source_location))
-        draft_location = as_draft(source_location)
-        if draft_location.category in DIRECT_ONLY_CATEGORIES:
-            raise InvalidVersionError(source_location)
-        original['_id'] = draft_location.dict()
-        try:
-            self.collection.insert(original)
-        except pymongo.errors.DuplicateKeyError:
-            raise DuplicateItemError(original['_id'])
-
-        self.refresh_cached_metadata_inheritance_tree(draft_location)
-        self.fire_updated_modulestore_signal(get_course_id_no_run(draft_location), draft_location)
-
-        return self._load_items([original])[0]
+        if Location(location).category in DIRECT_ONLY_CATEGORIES:
+            raise InvalidVersionError(location)
+        return wrap_draft(super(DraftModuleStore, self).clone_item(source, as_draft(location)))
 
     def update_item(self, location, data, allow_not_found=False):
         """
@@ -168,7 +140,7 @@ class DraftModuleStore(MongoModuleStore):
         try:
             draft_item = self.get_item(location)
             if not getattr(draft_item, 'is_draft', False):
-                self.convert_to_draft(location)
+                self.clone_item(location, draft_loc)
         except ItemNotFoundError, e:
             if not allow_not_found:
                 raise e
@@ -186,7 +158,7 @@ class DraftModuleStore(MongoModuleStore):
         draft_loc = as_draft(location)
         draft_item = self.get_item(location)
         if not getattr(draft_item, 'is_draft', False):
-            self.convert_to_draft(location)
+            self.clone_item(location, draft_loc)
 
         return super(DraftModuleStore, self).update_children(draft_loc, children)
 
@@ -202,7 +174,7 @@ class DraftModuleStore(MongoModuleStore):
         draft_item = self.get_item(location)
 
         if not getattr(draft_item, 'is_draft', False):
-            self.convert_to_draft(location)
+            self.clone_item(location, draft_loc)
 
         if 'is_draft' in metadata:
             del metadata['is_draft']
@@ -246,7 +218,9 @@ class DraftModuleStore(MongoModuleStore):
         """
         Turn the published version into a draft, removing the published version
         """
-        self.convert_to_draft(location)
+        if Location(location).category in DIRECT_ONLY_CATEGORIES:
+            raise InvalidVersionError(location)
+        super(DraftModuleStore, self).clone_item(location, as_draft(location))
         super(DraftModuleStore, self).delete_item(location)
 
     def _query_children_for_cache_children(self, items):
